refactor: update dashboard, sidebar to use loading component
