+++ jupyterlab_caip_optimizer/src/components/suggest_trials/create_trial.tsx
@@ -14,7 +14,7 @@ import {
   Typography,
 } from '@material-ui/core';
 import { prettifyStudyName } from '../../service/optimizer';
-import { Loading } from './loading';
+import { Loading } from '../loading';
 import { MetricInputs } from './metric_inputs';
 import { useDispatch } from 'react-redux';
 import { MetricsInputs } from '.';
